+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## 0.6.0 [unreleased]
 
 ### Features
+
 1. [#89](https://github.com/InfluxCommunity/influxdb3-python/pull/89): Use `datetime.fromisoformat` over `dateutil.parse` in Python 3.11+
+1. [#92](https://github.com/InfluxCommunity/influxdb3-python/pull/92): Update `user-agent` header value to `influxdb3-python/{VERSION}` and add it to queries as well. 
 
 ## 0.5.0 [2024-05-17]